+++ lib/spree_square.rb
@@ -2,6 +2,7 @@ require 'spree'
 require 'spree_square/engine'
 require 'spree_square/version'
 require 'spree_square/configuration'
+require 'spree_square/square_sdk_false_values'
 
 module SpreeSquare
   mattr_accessor :queue